pyproject.toml
[tool.poetry]
name = "puffcomqtt"
version = "0.1.0"
description = ""
authors = ["Martin George <marty@dingopaws.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
aiomqtt = "^2.0.1"
pyyaml = "^6.0.1"
bleak = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
